requirements.txt
pytorch-lightning<2
omegaconf==2.2.3
nerfacc==0.3.3
matplotlib
opencv-python
imageio
imageio-ffmpeg
scipy
PyMCubes
pyransac3d
torch_efficient_distloss
tensorboard
lpips
torch_scatter
prettytable
trimesh
